Finish Version 2.1.2
1. update some default value
2. transfer it with a new domain [remove bycompose]
3. optimize some logic with xor instead of pass a new params.
