+++ .gitignore
@@ -34,3 +34,4 @@ google-services.json
 
 app/release
 app/.DS_Store
+.DS_Store
