+++ app/src/main/java/com/apollo/timeflow/TimeFlowApplication.kt
@@ -1,0 +1,7 @@
+package com.apollo.timeflow
+
+import android.app.Application
+import dagger.hilt.android.HiltAndroidApp
+
+@HiltAndroidApp
+class TimeFlowApplication : Application()
