--- app/src/main/java/com/apollo/timeflow/bycompose/CoroutinesScopesModule.kt
@@ -1,23 +1,0 @@
-package com.apollo.timeflow.bycompose
-
-import dagger.Module
-import dagger.Provides
-import dagger.hilt.InstallIn
-import dagger.hilt.components.SingletonComponent
-import kotlinx.coroutines.CoroutineScope
-import kotlinx.coroutines.Dispatchers
-import kotlinx.coroutines.SupervisorJob
-import javax.inject.Singleton
-import kotlin.coroutines.CoroutineContext
-
-@InstallIn(SingletonComponent::class)
-@Module
-object CoroutinesScopesModule {
-    @Singleton
-    @Provides
-    fun providesCoroutineScope(): CoroutineScope = CoroutineScope(SupervisorJob() + Dispatchers.IO)
-
-    @Singleton
-    @Provides
-    fun providesCoroutineContext(): CoroutineContext = CoroutineScope(SupervisorJob() + Dispatchers.IO).coroutineContext
-}
