--- app/src/main/java/com/apollo/timeflow/bycompose/TimeFlowApplication.kt
@@ -1,7 +1,0 @@
-package com.apollo.timeflow.bycompose
-
-import android.app.Application
-import dagger.hilt.android.HiltAndroidApp
-
-@HiltAndroidApp
-class TimeFlowApplication : Application()
