--- app/src/main/java/com/apollo/timeflow/bycompose/broadcast/DateBroadcast.kt
@@ -1,11 +1,0 @@
-package com.apollo.timeflow.bycompose.broadcast
-
-import android.content.BroadcastReceiver
-import android.content.Context
-import android.content.Intent
-
-class DateBroadcast(private val updateDate: () -> Unit): BroadcastReceiver() {
-    override fun onReceive(context: Context?, intent: Intent?) {
-        updateDate.invoke()
-    }
-}
