--- app/src/main/java/com/apollo/timeflow/bycompose/ui/theme/Color.kt
@@ -1,10 +1,0 @@
-package com.apollo.timeflow.bycompose.ui.theme
-
-import androidx.compose.ui.graphics.Color
-
-
-val m3_grey50 = Color(0xFFFAFAFA)
-val m3_grey_900 = Color(0xFF212121)
-
-val m3_white = Color.White
-val m3_black = Color.Black
