--- app/src/main/java/com/apollo/timeflow/bycompose/uistate/DateUIState.kt
@@ -1,6 +1,0 @@
-package com.apollo.timeflow.bycompose.uistate
-
-data class DateUIState(
-    val showOrHide: Boolean = false,
-    val currentDate: String = "",
-)
