--- app/src/main/java/com/apollo/timeflow/bycompose/utils/TimeFormat.kt
@@ -1,6 +1,0 @@
-package com.apollo.timeflow.bycompose.utils
-
-enum class TimeFormat {
-    Base12,
-    Base24
-}
