+++ app/src/main/java/com/apollo/timeflow/ui/theme/Color.kt
@@ -1,0 +1,10 @@
+package com.apollo.timeflow.ui.theme
+
+import androidx.compose.ui.graphics.Color
+
+
+val m3_grey50 = Color(0xFFFAFAFA)
+val m3_grey_900 = Color(0xFF212121)
+
+val m3_white = Color.White
+val m3_black = Color.Black
