+++ app/src/main/java/com/apollo/timeflow/uistate/DateUIState.kt
@@ -1,0 +1,6 @@
+package com.apollo.timeflow.uistate
+
+data class DateUIState(
+    val showOrHide: Boolean = false,
+    val currentDate: String = "",
+)
