+++ app/src/main/java/com/apollo/timeflow/uistate/TimeUIState.kt
@@ -1,0 +1,11 @@
+package com.apollo.timeflow.uistate
+
+import androidx.annotation.StringRes
+
+data class TimeUIState(
+    val hoursLeft: Int,
+    val hoursRight: Int,
+    val minutesLeft: Int,
+    val minutesRight: Int,
+    @StringRes val amOrPM: Int,
+)
