+++ app/src/main/java/com/apollo/timeflow/utils/LogUtils.kt
@@ -1,0 +1,7 @@
+package com.apollo.timeflow.utils
+
+import android.util.Log
+
+fun String.logd() {
+    Log.d("TImeFlowLog", this)
+}
