+++ app/src/main/java/com/apollo/timeflow/utils/TimeFormat.kt
@@ -1,0 +1,6 @@
+package com.apollo.timeflow.utils
+
+enum class TimeFormat {
+    Base12,
+    Base24
+}
